Extract isSessionExpired
